Extract Plated into standalone plated and plated-instances packages
