+++ cabal.project
@@ -1,3 +1,5 @@
 packages: .
           ./examples
           ./lens-properties
+          ./plated
+          ./plated-instances
