+++ lens.cabal
@@ -198,6 +198,8 @@ library
     kan-extensions                >= 5        && < 6,
     mtl                           >= 2.2.1    && < 2.4,
     parallel                      >= 3.2.1.0  && < 3.4,
+    plated                        >= 0.1      && < 0.2,
+    plated-instances              >= 0.1      && < 0.2,
     profunctors                   >= 5.5.2    && < 6,
     reflection                    >= 2.1      && < 3,
     semigroupoids                 >= 5.0.1    && < 7,
