+++ plated-instances/plated-instances.cabal
@@ -1,0 +1,54 @@
+name:          plated-instances
+category:      Data, Generics
+version:       0.1
+license:       BSD2
+cabal-version: 1.18
+license-file:  LICENSE
+author:        Edward A. Kmett
+maintainer:    Edward A. Kmett <ekmett@gmail.com>
+stability:     provisional
+homepage:      http://github.com/ekmett/lens/
+bug-reports:   http://github.com/ekmett/lens/issues
+copyright:     Copyright (C) 2012-2016 Edward A. Kmett
+synopsis:      Plated instances for the free and comonad type families
+description:
+  Orphan @Plated@ instances (for the class in the @plated@ package) for the
+  @free@ and @comonad@ type families: @Free@, @FreeT@, @F@, @Cofree@, and
+  @CofreeT@. Kept separate so that depending on the @Plated@ class itself does
+  not pull in @free@/@comonad@. Importing @Data.Plated.Instances@ brings the
+  instances into scope.
+build-type:    Simple
+tested-with:   GHC == 8.0.2
+             , GHC == 8.2.2
+             , GHC == 8.4.4
+             , GHC == 8.6.5
+             , GHC == 8.8.4
+             , GHC == 8.10.7
+             , GHC == 9.0.2
+             , GHC == 9.2.8
+             , GHC == 9.4.8
+             , GHC == 9.6.7
+             , GHC == 9.8.4
+             , GHC == 9.10.3
+             , GHC == 9.12.2
+             , GHC == 9.14.1
+extra-source-files:
+  CHANGELOG.markdown
+
+source-repository head
+  type: git
+  location: https://github.com/ekmett/lens.git
+
+library
+  build-depends:
+    base       >= 4.9      && < 5,
+    comonad    >= 5.0.7    && < 6,
+    free       >= 5.1.5    && < 6,
+    plated     >= 0.1      && < 0.2
+
+  exposed-modules:
+    Data.Plated.Instances
+
+  hs-source-dirs: src
+  ghc-options: -Wall -Wno-orphans
+  default-language: Haskell2010
